Move into different files and refactor to use different line styles
